+++ wpipe_steps/infrastructure/__init__.py
@@ -1,3 +1,15 @@
 from .s3 import S3BucketUploadStep
+from .docker import DockerContainerStep
+from .kubernetes import KubernetesPodCheckStep
+from .terraform import TerraformApplyStep
+from .proxmox import ProxmoxVMStep
+from .digitalocean import DigitalOceanDropletStep
 
-__all__ = ["S3BucketUploadStep"]
+__all__ = [
+    "S3BucketUploadStep",
+    "DockerContainerStep",
+    "KubernetesPodCheckStep",
+    "TerraformApplyStep",
+    "ProxmoxVMStep",
+    "DigitalOceanDropletStep"
+]